Add 'mode' attribute to AllReduce and ReduceScatter.
This is an enum attribute which holds a CollectiveOpGroupMode. This enum determines how IDs in the replica_groups attribute are interpreted. Currently it is derived from the use_global_device_ids attribute and whether or not the channel_id attribute is present. The value of channel_id is not used in determine the mode, and is unused in all of XLA except for rare cases on TPUs.

The long-term plan is to instead directly have instructions hold the CollectiveOpGroupMode, remove the use_global_device_ids attribute, and no longer use channel_id to determine the mode. An instruction will then only have a channel_id in the rare cases where it's value (and not just its presence) is used.

This change is the first step of this plan. It adds the mode attribute to AllReduce and ReduceScatter. The constructor of such instructions takes an optional<CollectiveOpGroupMode> defaulting to nullopt, which causes the mode to be derived from channel_id and use_global_device_ids. The instructions store a non-optional CollectiveOpGroupMode. The HLO verifier will complain if the mode does not match channel_id and use_global_device_ids. The mode is optional in HLO text, and if not present, will be derived from channel_id and use_global_device_ids.

CollectiveOpGroupMode is moved to its own file to avoid a circular dependency between collective_ops_utils.h and hlo_instructions.h. I created a duplicate CollectiveOpGroupModeProto enum, since we need a proto form for HloInstructionProto. Eventually we should get rid of the C++ enum and just use the proto enum.

After this change, the next step is to add the attribute to other collectives. Then, change usages of functions like CreateAllReduce to pass the mode, and change passes and other code to directly use the instruction's mode rather than calculate it from channel_id and use_global_device IDs. The final step is to remove use_global_device_ids and to make channel_id unused except in the cases where its integer value, not just the existence of a nonzero int, is used.

PiperOrigin-RevId: 777811427